// model/accountStatsSpaceReservation.ts
/**
 * ManageSpace Lease API
 * ManageSpace Lease API Documentation
 *
 * The version of the OpenAPI document: 0.0.12-dev
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class AccountStatsSpaceReservation {
    /**
    * The unique identifier for the account
    */
    'accountId': string;
    /**
    * The name of the account
    */
    'accountName': string;
    /**
    * The email of the account
    */
    'email': string;
    /**
    * The phone number of the account
    */
    'phoneNumber': string;
    /**
    * The count of reserved spaces for an account.
    */
    'reservedCount': number;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "accountId",
            "baseName": "accountId",
            "type": "string"
        },
        {
            "name": "accountName",
            "baseName": "accountName",
            "type": "string"
        },
        {
            "name": "email",
            "baseName": "email",
            "type": "string"
        },
        {
            "name": "phoneNumber",
            "baseName": "phoneNumber",
            "type": "string"
        },
        {
            "name": "reservedCount",
            "baseName": "reservedCount",
            "type": "number"
        }    ];

    static getAttributeTypeMap() {
        return AccountStatsSpaceReservation.attributeTypeMap;
    }
}