// model/calculateProratedAmountRequest.ts
/**
 * ManageSpace Lease API
 * ManageSpace Lease API Documentation
 *
 * The version of the OpenAPI document: 0.0.11-dev
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class CalculateProratedAmountRequest {
    /**
    * Future date for proration.
    */
    'futureDate': Date;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "futureDate",
            "baseName": "futureDate",
            "type": "Date"
        }    ];

    static getAttributeTypeMap() {
        return CalculateProratedAmountRequest.attributeTypeMap;
    }
}