// model/calculateProratedAmountResponse.ts
/**
 * ManageSpace Lease API
 * ManageSpace Lease API Documentation
 *
 * The version of the OpenAPI document: 0.0.12-dev
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class CalculateProratedAmountResponse {
    /**
    * A positive number representing amount due or a negative amount representing credit.
    */
    'amount': number;
    /**
    * Payment provider description for charges.
    */
    'description': string;
    /**
    * Payment provider description for charges.
    */
    'detailedDescription': string;
    /**
    * Representing the beginning of the period for which the proration is being calculated.
    */
    'prorationStartDate': Date;
    /**
    * Representing the end of the period for which the proration is being calculated.
    */
    'prorationEndDate': Date;
    /**
    * Date representing the end of the current billing period.
    */
    'originalBillingCycleAnchor': Date;
    /**
    * Date representing the end of the current billing period.
    */
    'paidThrough': Date;
    /**
    * Adjusted date used for proration calculation.
    */
    'adjustedFutureDate': Date;
    /**
    * Amount in centes representing the price associated with the current subscription plan.
    */
    'subscriptionAmount': number;
    /**
    * Payment provider customer id
    */
    'customerId': string;
    /**
    * Payment provider subscription id
    */
    'subscriptionId': string;
    /**
    * A positive number representing amount due or a negative amount representing credit.
    */
    'daysBetween': number;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "amount",
            "baseName": "amount",
            "type": "number"
        },
        {
            "name": "description",
            "baseName": "description",
            "type": "string"
        },
        {
            "name": "detailedDescription",
            "baseName": "detailedDescription",
            "type": "string"
        },
        {
            "name": "prorationStartDate",
            "baseName": "prorationStartDate",
            "type": "Date"
        },
        {
            "name": "prorationEndDate",
            "baseName": "prorationEndDate",
            "type": "Date"
        },
        {
            "name": "originalBillingCycleAnchor",
            "baseName": "originalBillingCycleAnchor",
            "type": "Date"
        },
        {
            "name": "paidThrough",
            "baseName": "paidThrough",
            "type": "Date"
        },
        {
            "name": "adjustedFutureDate",
            "baseName": "adjustedFutureDate",
            "type": "Date"
        },
        {
            "name": "subscriptionAmount",
            "baseName": "subscriptionAmount",
            "type": "number"
        },
        {
            "name": "customerId",
            "baseName": "customerId",
            "type": "string"
        },
        {
            "name": "subscriptionId",
            "baseName": "subscriptionId",
            "type": "string"
        },
        {
            "name": "daysBetween",
            "baseName": "daysBetween",
            "type": "number"
        }    ];

    static getAttributeTypeMap() {
        return CalculateProratedAmountResponse.attributeTypeMap;
    }
}